% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{colour_options}
\alias{colour_options}
\title{Display the base R colour options.}
\usage{
colour_options(print_to_pdf = FALSE, pdf_name = "base_r_colour_options.pdf")
}
\arguments{
\item{print_to_pdf}{IF TRUE, output is exported to a PDF in the working
directory instead of displayed in an R graphics device.}

\item{pdf_name}{The name of the PDF file. Default is
"base_r_colour_options.pdf". Must end in .pdf. Ignored if print_to_pdf =
FALSE.}
}
\description{
display the base R colour options in the active graphics
plotting device or printed to a PDF file. This makes it easier to see what
the color options are in conjunction with their R labels to make it easier
to specify custom colours for plots.
}
\examples{
\dontrun{
#default settings. Prints to the active graphics device.
colour_options(print_to_pdf = FALSE) #the default
colour_options(print_to_pdf = TRUE) #print to a single page PDF file in your working directory.
}
}
\references{
Code was adapted from \url{http://bc.bojanorama.pl/2013/04/r-color-reference-sheet/}
}
\author{
Craig P. Hutton, \email{Craig.Hutton@gov.bc.ca}
}
