% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{consum}
\alias{consum}
\title{consecutive summation}
\usage{
consum(x, skip_na = FALSE)
}
\arguments{
\item{x}{the logical or binary vector to summed.}

\item{skip_na}{Should consecutive counting resume or reset after encountering
a missing value (NA)? Default is FALSE.}
}
\value{
A numeric vector of the same length as x with the consecutive total.
}
\description{
Calculate a rolling sum of consecutive TRUE values for a logical
vector or "1" values for a binary vector. The total will continue
increasing as it moves along the vector and encounters only 1/TRUE, then
will reset after encountering a 0/FALSE value. When working with monthly
service utilization data with rows organized by ID and month, this is
useful for estimating things like how many consecutive months someone used
a service.
}
\examples{

x <- c(rep(1, 8), rep(0, 5), rep(1, 4), rep(NA_real_, 2), rep(1, 6))

consum(x)

consum(x, skip_na = TRUE)


}
\seealso{
\code{\link{match}}, \code{\link{Negate}}
}
\author{
Craig P. Hutton, \email{Craig.Hutton@gov.bc.ca}
}
