% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{counts_all}
\alias{counts_all}
\title{Obtain the counts for unique values of all variables in a data frame.}
\usage{
counts_all(data, n = "all", order = c("d", "a", "i"), sep = "_", na.rm = TRUE)
}
\arguments{
\item{data}{A data frame (required).}

\item{n}{The number of unique values you want frequency counts for. Default is "all".}

\item{order}{"d" for descending/decreasing order. "a" or "i" for
ascending/increasing order.}

\item{sep}{A character string to use to separate unique values from their counts
("_" by default).}

\item{na.rm}{Should missing values be omitted (TRUE/FALSE)?}
}
\value{
A list of character vectors of the unique value frequency counts for
each variable of the input data frame sorted in the chosen order. Return
values are structured as "value_count", where the "_" portion takes on the
value of the sep argument. Returning a character vector makes subsequent
manipulation with stringr and other tidyverse tools fairly easily.
}
\description{
\code{counts_all} is an extension of \code{\link{counts}} that
returns the most or least common unique value(s) for each column vector of
a data frame. Also useful for identifying data entry errors or rare cases.
For complex use cases see \code{\link{describe_all}}.
}
\examples{
#using a numeric vector
data <- data(mtcars)

#all unique values of all variables in \code{data} in order of descending frequency
counts_all(data)

#the most common values of all variables in \code{data}
counts_all(data, n = 1)

#the top 5 most common unique values for all variables in  \code{data}
counts_all(data, n = 5)

}
\seealso{
\code{\link{table}}, \code{\link{sort}}, \code{\link[purrr]{map}}
}
\author{
Craig P. Hutton, \email{Craig.Hutton@bov.bc.ca}
}
