% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{counts_tb}
\alias{counts_tb}
\title{Obtain the top and bottom "n" counts for unique values of a vector.}
\usage{
counts_tb(y, n = 10, sep = "_", na.rm = TRUE)
}
\arguments{
\item{y}{A vector.}

\item{n}{The number of top & bottom unique values you want frequency counts
for. Default is 10 (fewer than "n" will be shown if there aren't "n" unique
values).}

\item{sep}{This only needs to be modified if some values to be counted
contain an underscore, in which case you should change it to a character
string that is not present in any of the values of y.}

\item{na.rm}{Should missing values be omitted (TRUE/FALSE)?}
}
\value{
A list of data frames of the top and bottom counts for each variable
of the input data frame. Return value columns are "top_v" = top value,
"top_n" = count of the top value in the same row of the adjacent top_v
column, "bot_v" = bottom value, & "bot_n" = count of the bottom value in
the same row of the adjacent bot_v column.
}
\description{
\code{counts_tb} is a convenience extension of
\code{\link{counts}} that returns the most AND least common "n" unique
value(s) a vector. This is useful for identifying data entry errors or rare
cases. For complex use cases see \code{\link{describe}}.
}
\examples{
#using the mtcars data
data(mtcars)

counts_tb(mtcars$cyl) #top & bottom values

}
\seealso{
\code{\link{counts_tb_all}}, \code{\link{counts_all}}, \code{\link{counts}}
}
\author{
Craig P. Hutton, \email{Craig.Hutton@gov.bc.ca}
}
