% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Obtain a descriptive summary of a variable.}
\usage{
describe(data, y, ..., n = "all", interactive = FALSE, paging = TRUE,
  caption = NULL)
}
\arguments{
\item{data}{A data frame or tibble containing the dependent measure "Y" and
any grouping variables.}

\item{y}{The numeric variable for which you wish to obtain a descriptive
summary and normality assessment}

\item{...}{Any number of grouping variables (also present in the data source)
to use for subsetting, separated by commas (e.g. \code{group_var1,
  group_var2})}

\item{n}{The number of unique values you want frequency counts for (default =
5). Ignored if y is numeric. If you want all unique values or the unique
value counts for a numeric variable, use code{\link{mcvals}} instead.}

\item{interactive}{Determines whether a data frame or an interactive html
table with options for editing and exporting and in line box plots and
histograms produced by \code{\link[sparkline]{sparkline}} and
\code{\link[DT]{datatable}} are returned.}

\item{paging}{Set to F to disable paging of an interactive table. Ignored if
interactive = F.}

\item{caption}{Add a caption/title to the table (interactive mode only).}
}
\value{
A vector of summary statistics if y is numeric, or sample size info
and the most common n values of y if it is non-numeric, split into
group(s).

For all input variables, the following are returned:

cases = the total number of cases

n = number of complete cases

na = number of missing values

p_na = percentage of total cases with missing values

These additional measures are provided for numeric variables:

mean = the mean of y

mode = the mode AKA most common value of y

sd = standard deviation of y

se = standard error of the mean of y

p0 = the minimum AKA the 0th percentile of y

p25 = the 25th percentile of y

p50 = the median AKA 50th percentile of y

p75 = the 75th percentile of y

p100 = the maximum AKA the 100th percentile of y

IQR = the interquartile range of y (75th percentile - 25th percentile)

hist = histogram of y (interactive mode only)

box = boxplot of y (interactive mode only)

For non-numeric variables, the top n (default = 5) most common unique
values and their frequencies are also returned, e.g. Val_1 = most common
value, Ct_1 = # of occurences of the most common value, etc.
}
\description{
Obtain a useful array of common summary statistics for a numeric
variable/vector, or the counts of the most frequent unique values of a
non-numeric variable, within a data frame with or without first separating
it into groups based on any number of grouping variables. Also
pipe-friendly (data is the 1st argument). While other similar functions
exist in other packages (e.g. \code{\link[psych]{describeBy}}), the use of
tidy evaluation using \code{\link[dplyr]{dplyr}} pipes makes it easy to subset
using an arbitrary number of grouping variables. For exploratory or
publication/presentation purposes, further customization is made easier
using the interactive mode.
}
\examples{
data(gapminder, package = "gapminder") #load the gapminder life expectancy data

#without piping:
describe(data = gapminder, y = lifeExp) #no grouping variables
describe(gapminder, lifeExp, continent) #one grouping variable
describe(gapminder, lifeExp, continent, country) #two grouping variables

#with piping:
library(magrittr)

gapminder \%>\% describe(lifeExp) #no grouping variables
gapminder \%>\% describe(lifeExp, continent) #one grouping variables
gapminder \%>\% describe(lifeExp, continent, country) #two grouping variables

#interactive version
\dontrun{
gapminder \%>\% describe(lifeExp, continent, interactive = T)
}


}
\references{
Altman, D. G., & Bland, J. M. (2005). Standard deviations and standard
errors. Bmj, 331(7521), 903.

Bulmer, M. G. (1979). Principles of statistics. Courier Corporation.
}
\seealso{
\code{\link[base]{mean}}, \code{\link[stats]{sd}}, \code{\link[elucidate]{se}},
\code{\link[stats]{quantile}}, \code{\link[stats]{IQR}},
\code{\link[sparkline]{sparkline}}, \code{\link[DT]{datatable}}
}
\author{
Craig P. Hutton, \email{craig.hutton@gmail.com}
}
