% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_all.R
\name{describe_all}
\alias{describe_all}
\title{Obtain a descriptive summary of all variables in a data frame or tibble.}
\usage{
describe_all(data, ..., class = "all", interactive = F, n = 5,
  paging = T, caption = NULL)
}
\arguments{
\item{data}{A data frame or tibble containing the at least one numerical
variable and any grouping variables.}

\item{...}{Any number of grouping variables (also present in the data source)
to use for subsetting, separated by commas (e.g. \code{group_var1,
  group_var2})}

\item{class}{Obtain a descriptive summary of numeric variables (class =
"numeric"), non-numeric variables (class = "other"), or all variables
(class = "all"; the default). If all is requested, the output will be a
list.}

\item{interactive}{Determines whether a data frame or an interactive html
table with in line box plots and histograms produced by
\code{\link[sparkline]{sparkline}} and \code{\link[DT]{datatable}} are
returned with options for editing and exporting. Since
\code{\link[DT]{datatable}} requires 2-dimensional input, interactive
tables are only generated if class = "numeric", class = "other", or if
class = "all" but all variables in the dataset are either "numeric" or
"non-numeric", but not a mixture.}

\item{n}{The number of unique values you want frequency counts of for the
non-numeric variables in the data object (default = 5). Ignored if class =
"numeric". If you want all unique values or the unique value counts for a
numeric variable, use code{\link[elucidate]{mcvals_all}} instead.}

\item{paging}{Set to F to disable paging of an interactive table (interactive
mode only).}

\item{caption}{Add a caption/title to the table (interactive mode only).}
}
\value{
A vector of summary statistics if y is numeric, or sample size info
and the most common n values of y if it is non-numeric, split into
group(s).

For all input variables, the following are returned:

cases = the total number of cases

n = number of complete cases

na = number of missing values

p_na = percentage of total cases with missing values

These additional measures are provided for numeric variables:

mean = the mean of y

mode = the mode AKA most common value of y

sd = standard deviation of y

se = standard error of the mean of y

p0 = the minimum AKA the 0th percentile of y

p25 = the 25th percentile of y

p50 = the median AKA 50th percentile of y

p75 = the 75th percentile of y

p100 = the maximum AKA the 100th percentile of y

IQR = the interquartile range of y (75th percentile - 25th percentile)

hist = histogram of y (interactive mode only)

box = boxplot of y (interactive mode only)

For non-numeric variables, the top n (default = 5) most common unique
values and their frequencies are also returned, e.g. Val_1 = most common
value, Ct_1 = # of occurences of the most common value, etc.
}
\description{
Extends \code{\link[elucidate]{describe}} by looping/mapping it (see
\code{\link[purrr]{map}} for more info.) across all variables in a data
frame or tibble, with or without first separating it into groups based on
any number of grouping variables. Also pipe-friendly (data is the 1st
argument). While other similar functions exist in other packages (e.g.
\code{\link[skimr]{skimr}}), describe_all provides an expanded set of summary
statistics and adds dynamic functionality via the interactive output option
that provides inline boxplots in addition to histograms. To obtain
confidence intervals and skewness, kurtosis, and normality metrics for the
sample mean and median of each numeric variable use
\code{\link[elucidate]{describe_ci_all}} instead. To obtain frequency counts for the
unique values/levels of all variables, use \code{\link[elucidate]{mcvals_all}}
instead. For exploratory or publication/presentation purposes, further
customization is made easier using the interactive mode, the output of
which can easily be copied or exported in various formats (e.g. csv, excel,
pdf, image file).
}
\examples{
library(gapminder) #load the gapminder dataset

#without piping:
describe_all(data = gapminder) #no grouping variables

describe_all(gapminder, continent) #one grouping variable

\dontrun{
#with piping:
library(magrittr)

gapminder \%>\% describe_all() #no grouping variables

gapminder \%>\% describe_all(continent) #one grouping variable

gapminder \%>\% describe_all(continent, country, class = "numeric") #two grouping variables

#arrange the output by a grouping factor instead of measurement variables using dplyr::arrange

gapminder \%>\% describe_all(continent) \%>\% dplyr::arrange(continent)

#filter the output using dplyr::filter to more easily compare groups on a subset of measures

gapminder \%>\%
describe_all(continent) \%>\%
dplyr::filter(variable == "lifeExp" | variable == "pop") #examine just life expectancy and populaton

#interactive version

gapminder \%>\% describe_all(continent, class = "numeric",interactive = T)
}

}
\seealso{
\code{\link[elucidate]{describe}}
}
\author{
Craig P. Hutton, \email{craig.hutton@gmail.com}
}
