% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_ci_all.R
\name{describe_ci_all}
\alias{describe_ci_all}
\title{Confidence intervals for the mean and/or median of all numeric variables in a
data frame}
\usage{
describe_ci_all(data, ..., ci_level = 0.95, mode = "p",
  sw_alpha = 0.05, replicates = 2000, interactive = FALSE,
  paging = TRUE, caption = NULL)
}
\arguments{
\item{data}{A data frame or tibble containing the dependent measure "Y" and
any grouping variables.}

\item{...}{Any number of grouping variables (also present in the data source)
to use for subsetting, separated by commas (e.g. \code{group_var1,
  group_var2})}

\item{ci_level}{The confidence level to use for constructing confidence intervals.
Default is set to \code{CI_level = 0.95} for 95 percent CIs.}

\item{mode}{Specify whether you want parametric (normal) CIs, based on the
mean (mode = "p"), non-parametric boostrapped CIs (mode = "np") based on
the median, or both (mode = "both"). For the parametric mode, only the
sample size, mean, normality assessments, skewness, kurtosis, and CIs based
on the normal distribution are provided. For the non-parametric mode, only
sample size info, the median, IQR, and bootstrapped CIs for the median are
provided. This is done to minimize processing time for large datasets so
that you can first check if the data are normally distributed (and whether
or not normal CIs based on the mean are therefore useful) before investing
the extra time it takes to get bootstrapped CIs. For all
results, use mode = "both".}

\item{sw_alpha}{The alpha-level to use to evaluate the null hypothesis for the
Shaprio-Wilk test. Default is set to \code{alpha = 0.05}.}

\item{replicates}{The number of bootstrap replicates to use for the BCa CIs
of the median (in non-parametric & combined modes). Default is 2,000, as
recommended by Efron & Tibshirani (1993). For publications, or if you need
more precise estimates, more replications (e.g. >= 5,000) are recommended.
N.B. more replications will of course take longer to run. If you get the
error: "estimated adjustment 'a' is NA" then try again with more
replications.}

\item{interactive}{Determines whether a data frame or an interactive html
table with options for editing and exporting and in line box plots and
histograms produced by \code{\link[sparkline]{sparkline}} and
\code{\link[DT]{datatable}} are returned.}

\item{paging}{Set to F to disable paging of an interactive table. Ignored if
interactive = F.}

\item{caption}{Add a caption/title to the table (interactive mode only).}
}
\value{
A vector of summary statistics and distribution test results, split
into group(s):

n = number of complete cases

na = number of missing values

mean = the mean of y

sd = standard deviation of y

mean_LB = lower bound of the confidence interval for the mean (assuming
normality)

mean_UB = upper bound of the confidence interval for the mean (assuming
normality)

p50 = the median AKA 50th percentile of y

IQR = the interquartile range of y = 75th percentile - 25th percentile

p50_LB = lower bound of the boostrapped confidence interval for the median
(no distributional assumptions)

p50_UB = upper bound of the bootsrapped confidence interval for the median
(no distributional assumptions)

skew = skewness of the distribution of y.

kurt = kurtosis of the distribution of y.

SW_W = the Shaprio-Wilk test statistic "W" (see
\code{\link[stats]{shapiro.test}}) comparing the distribution of y to a
normal distribution. The SW test requires an n between 3-3,000 to produce
informative/useful results and will consequently return "NA" if the sample
size falls outside of this range. If an n > 3,000 then normality is best
evaluated using graphical methods, e.g. the box plots and histograms
provided in the interactive mode (or a qqplot).

SW_p = the Shaprio-Wilk test p-value. Returns NA if n < 3 or > 3,000.

normal = whether or not the normality assumption is valid based on the
obtained p-value and chosen alpha level (default is 0.05). If the data are
non-normally distributed use the percentile(median)-based non-parametric
statistics, not the parametric ones. Requires n = 3-3,000.

hist = histogram of y (interactive mode only)

box = boxplot of y (interactive mode only)
}
\description{
Extends \code{\link[elucidate]{describe_ci}} by mapping it across all
numeric variables in a data frame or tibble, with or without first
separating it into groups based on any number of grouping variables. Also
pipe-friendly (data is the 1st argument). While other similar functions
exist in other packages (e.g. \code{\link[skimr]{skimr}}), describe_ci_all
provides an easier way to get groupwise confidence intervals and adds
dynamic functionality via the interactive output option that provides
inline boxplots in addition to histograms to facilitate rapid data
exploration. To obtain frequency counts for the unique values or levels of
all non-numeric (& numeric) variables, use \code{\link[elucidate]{mcvals_all}} instead
(depending on whether you want the results sorted by most or least frequent
unique values).
}
\examples{
library(gapminder) #load the gapminder dataset

#without piping:
describe_ci_all(data = gapminder) #no grouping variables

describe_ci_all(gapminder, continent) #one grouping variable

describe_ci_all(gapminder, continent, country) #two grouping variables

\dontrun{
#with piping:
library(magrittr)

gapminder \%>\% describe_ci_all() #no grouping variables

gapminder \%>\% describe_ci_all(continent) #one grouping variable

gapminder \%>\% describe_ci_all(continent, country) #two grouping variables

#arrange the output by a grouping factor instead of measurement variables using dplyr::arrange

gapminder \%>\% describe_ci_all(continent) \%>\% dplyr::arrange(continent)

#adjust the alpha level for the Shaprio-Wilk test
gapminder \%>\% describe_ci_all(SW_alpha = 0.1)
gapminder \%>\% describe_ci_all(SW_alpha = 0.01)

#adjust the confidence level for confidence intervals
gapminder \%>\% describe_ci_all(ci_level = 0.90)

#adjusting the confidence level affects both mean and median-derived ci_alls
gapminder \%>\% describe_ci_all(ci_level = 0.8, mode = "both")

#obtain non-parametric ci_alls based on the median only
gapminder \%>\% describe_ci_all(mode = "np")
gapminder \%>\% describe_ci_all(country, mode = "np")

#adjust the number of replicates to use in constructing the BCa ci_alls for the
#confidence interval of the median (note that this will affect run time)
gapminder \%>\% describe_ci_all(mode = "np", replicates = 5000)

#set interactive = TRUE or interactive = T for the interactive output instead of a tibble
gapminder \%>\% describe_ci_all(continent, interactive = T)
}

}
\references{
Altman, D. G., & Bland, J. M. (2005). Standard deviations and standard
errors. Bmj, 331(7521), 903.

Efron, B. (1987). Better bootstrap confidence intervals. Journal of the
American statistical Association, 82(397), 171-185.

Efron, B., & Tibshirani, R. J. (1993). An introduction to the bootstrap. New
York: Chapman & Hall.

Shapiro, S. S. and Wilk, M. B. (1965). "An analysis of variance test for
normality (complete samples)", Biometrika, 52, 3 and 4, pages 591-611.
}
\seealso{
\code{\link[elucidate]{describe_ci}}
}
\author{
Craig P. Hutton, \email{craig.hutton@gmail.com}
}
