% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe_na_all}
\alias{describe_na_all}
\title{Obtain a descriptive summary of missing values for all variables in a data frame.}
\usage{
describe_na_all(data, ..., digits = 4, output = c("dt", "tibble"))
}
\arguments{
\item{data}{A data frame or tibble.}

\item{...}{This special argument accepts any number of unquoted grouping
variable names (also present in the data source) to use for subsetting,
separated by commas, e.g. \verb{group_var1, group_var2}. Also accepts a
character vector of column names or index numbers, e.g. c("group_var1",
"group_var2") or c(1, 2), but not a mixture of formats in the same call. If
no column names are specified, all columns will be used.}

\item{digits}{This determines the number of digits used for rounding of
the "p_na" column in the output.}

\item{output}{Output type for each class of variables. dt" for data.table or
"tibble" for tibble.}
}
\value{
A tibble or data.table with the following columns in addition to any specified grouping variables:

\describe{
\item{cases}{the total number of cases}
\item{n}{number of complete cases}
\item{na}{the number of missing values}
\item{p_na}{the proportion of total cases with missing values}
}
}
\description{
This function extends \code{{describe_na}} by applying to it all
columns in a data frame using functional programming tools from the purrr
package (e.g. \code{\link[purrr]{map}}). To obtain a summary of missing
values for a single variable in a data frame use \code{\link{describe_na}}
instead. This function is a more efficient way of checking for missing
values than using \code{\link{describe_all}}, which calculates additional
summary statistics.
}
\examples{

describe_na_all(mtcars)

}
\seealso{
\code{\link{describe}}, \code{\link{describe_na}}, \code{\link{describe_all}}
}
\author{
Craig P. Hutton, \email{craig.hutton@gov.bc.ca}
}
