% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{dscr_all}
\alias{dscr_all}
\title{elucidate package internal function}
\usage{
dscr_all(data, class = "all", n = 5, digits = 3, type = 2,
  output = "dt", order = "d", na.rm = TRUE)
}
\arguments{
\item{data}{input vector (required).}

\item{class}{passed to \code{\link{describe}}}

\item{n}{passed to \code{\link{describe}}}

\item{digits}{passed to \code{\link{describe}}}

\item{type}{passed to \code{\link{describe}}}

\item{output}{passed to \code{\link{describe}}}

\item{order}{passed to \code{\link{describe}}}

\item{na.rm}{passed to \code{\link{describe}}}
}
\description{
\code{dscr_all} is an internal function that supports
\code{\link{describe_all}}.
}
\author{
Craig P. Hutton, \email{craig.hutton@gmail.com}
}
