% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copies.R
\name{dupes}
\alias{dupes}
\title{Check the number of duplicated rows in a data frame.}
\usage{
dupes(
  data,
  ...,
  keep_all_cols = TRUE,
  sort_by_copies = TRUE,
  order = c("d", "a", "i"),
  na_last = FALSE,
  output = c("same", "tibble", "dt", "data.frame")
)
}
\arguments{
\item{data}{a data frame, tibble, or data.table.}

\item{...}{This special argument accepts any number of unquoted column names
(also present in the data source) to use when searching for duplicates. If
no column names are specified, all columns will be used.}

\item{keep_all_cols}{If column names are specified using \code{...}, this allows
you to drop unspecified columns, similarly to the \code{.keep_all} argument for
`dplyr::distinct()``}

\item{sort_by_copies}{If TRUE (the default), sorts the results by the number
of copies, in order specified by the \code{order} argument.}

\item{order}{If sort_by_copies is set to TRUE, this controls whether the
results should be sorted in order of descending/decreasing = "d" (the
default) or ascending/increasing = "a" or "i" copy numbers.}

\item{na_last}{should rows of the specified columns with missing values be
listed below non-missing values (TRUE/FALSE)? Default is FALSE.}

\item{output}{"tibble" for tibble, "dt" for data.table, or "data.frame" for a
data frame. "same", the default option, returns the same format as the
input data.}
}
\value{
A subset of the input data frame consisting of duplicated rows that were
detected based on specified variables used to condition the search. A
message will also be printed to the console indicating whether or not
duplicates were detected. An \code{n_copies} column is appended specifying the
total number of copies of each row that were detected.
}
\description{
Checks a data frame for duplicated rows based on specified
variables to use for checking (via \code{...}) or all columns (if
unspecified).\code{dupes} is a convenience shortcut for \code{\link{copies}}
with the "filter" argument set to "dupes" and the "sort_by_copies" argument
set to TRUE by default. For greater flexibility in checking row copy
numbers or filtering for distinct rows, use \code{\link{copies}} instead.
\code{dupes} behaves similarly to \code{\link[janitor]{get_dupes}}) but is
substantially faster due to the use of \code{data.table} as a backend.
}
\examples{

# check for duplicates based on one variable, "g" in this case
dupes(pdata, g)

\dontrun{
dupes(pdata, high_low, g) #check based on 2 variables

# check based on all variables, i.e. fully duplicated rows
dupes(pdata)
}

}
\seealso{
\code{\link{copies}}, \code{\link[janitor]{get_dupes}}
}
\author{
Craig P. Hutton, \email{craig.hutton@gov.bc.ca}
}
