% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elucidate.R
\docType{package}
\name{elucidate}
\alias{elucidate}
\alias{elucidate-package}
\title{"elucidate: a collection of convenience functions to help researchers elucidate patterns in their data."}
\description{
"This package provides a collection of convenience functions to facilitate
exploratory data analysis by researchers. While many of the functions are
geared towards data analysis tasks typical in the behavioural and health
sciences (e.g. analysis of numerical variables between experimental groups),
they are more widely applicable. The package has also been designed with the
goal of making common analytical processes in R more accessible to (data)
science trainees by providing a consistent API for data analysis and
visualization. Note This package is currently in development and its
functionality will be improved over time."
}
