% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{inv_quantile}
\alias{inv_quantile}
\title{Get the quantile(s) corresponding to a vector of values.}
\usage{
inv_quantile(y, values, digits = NULL)
}
\arguments{
\item{y}{a numeric vector/variable.}

\item{values}{the values or vector of specific values in y for which you would like the to know the quantiles}

\item{digits}{determines the how many digits the result is rounded to}
}
\description{
\code{inv_quantile} returns the quantile(s) for a specific value or vector of values.
}
\examples{
y <- c(1:100)
inv_quantile(y = y, values = c(2, 25, 50, 75, 95), digits = 2)

}
\seealso{
\code{\link[stats]{ecdf}}, \code{\link{round}}
}
\author{
Craig P. Hutton, \email{Craig.Hutton@bov.bc.ca}
}
