% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{mcvals}
\alias{mcvals}
\title{Obtain the most common values or value (i.e. the mode) of a vector.}
\usage{
mcvals(data = NULL, y, ..., n = "all")
}
\arguments{
\item{data}{Data frame or tibble containing the column you want the most
common values for (optional).}

\item{y}{A vector/variable (required).}

\item{...}{Any number of grouping variables (also present in the data source)
to use for subsetting, separated by commas (e.g. \code{group_var1,
  group_var2}). If grouping variables are supplied, then data must also be.}

\item{n}{The number of unique values you want frequency counts for (in order
of descending count). Default is "all".}
}
\value{
A data frame of the most common values from the input vector sorted in
descending order. The names of the output, printed as the 1st row when the
output is called, are the unique input values of y. The 2nd row of the
output contains the counts of the identified unique values. If grouping
variables are specified, then the values and counts for each level of the
grouping variable are returned. See the \code{\link{table}} function
documentation for more information.
}
\description{
\code{mcvals} returns the most common unique value(s) of a
vector. Also useful for identifying data entry errors or rare cases.
}
\examples{
#using a numeric vector
y <- c(1, 1, 2, 3, 4, 5, 6, 2, 3, 4, 2, 9, 5, 6, 7, 23, 5, 6, 7, 8, 3, 4, 5, 6)

mcvals(y = y) #all unique values of y in order of descending frequency \\code{y}
mcvals(y = y, n = 1) #the most common value of \\code{y}
mcvals(y = y, n = 5) #the top 5 most common unique values of \\code{y}

#using a character vector
ch_vect <- c("a", "b", "c", "a", "b", "a")
mcvals(y = ch_vect)
mcvals(y = ch_vect, n = 1)
mcvals(y = ch_vect, n = 5)

#top 5 values of y split by a grouping variable
data(mtcars)
mcvals(mtcars, y = mpg, cyl, n = 5)

}
\seealso{
\code{\link{table}}, \code{\link{sort}},
\code{\link{length}}
}
\author{
Craig P. Hutton, \email{craig.hutton@gmail.com}
}
