% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ci.R
\name{mean_ci}
\alias{mean_ci}
\title{Obtain a (bootstrapped) confidence interval for the mean of a numeric vector.}
\usage{
mean_ci(y, replicates = 2000, ci_level = 0.95, ci_type = "norm",
  parallel = FALSE, cores = NULL, na.rm = TRUE)
}
\arguments{
\item{y}{A vector/variable (required).}

\item{replicates}{The number of bootstrap replicates to use. Default is
2,000, as recommended by Efron & Tibshirani (1993). For publications, or if
you need more precise estimates, more replications (e.g. >= 5,000) are
recommended. N.B. more replications will of course take longer to run. If
you get the error: "estimated adjustment 'a' is NA" when ci_type is set to
"bca" then try again with more replications.}

\item{ci_level}{The confidence level to use for constructing confidence intervals.
Default is set to \code{ci_level = 0.95} for 95 percent CIs.}

\item{ci_type}{The type of confidence intervals to calculate from the
bootstrap samples. Most of the options available in the underlying boot.ci
function are implemented (except for studentized intervals): "norm" for
calculation based on a theoretical normal distribution, "perc" for
percentile, "basic" for basic, and "bca" for bias-corrected and
accelerated. See \code{\link[boot]{boot.ci}} for details regarding options
other than "norm". Since the normal confidence intervals for the mean can
be directly calculation based quantiles from the theoretical gaussian
distribution this method is used for this unique case (CIs for the mean)
instead of bootstrapping when the ci_type is set to "norm" (the default),
since it is MUCH faster.}

\item{parallel}{set to TRUE if you want to use multiple cores or FALSE if you
don't (the default). Note that there is some processing overhead involved
when operating in parallel so speed gains may not be very noticable for
smaller samples (and may even take longer than sequential processing). Due
to the nature of the underlying parallelization architecture, performance
gains will likely be greater on non-Windows machines that can use the
"multicore" implementation instead of "snow". For obvious reasons this
option only works on machines with more than 1 logical processing core.}

\item{cores}{If parallel is set to TRUE, this determines the number of cores
to use. To see how many cores are available on your machine, use
parallel::detectCores()}

\item{na.rm}{should missing values be removed before attempting to calculate
the mean and confidence intervals? Default is TRUE.}
}
\description{
mean_ci returns confidence intervals for the mean of a numeric
vector. One might want to use bootstrapping to obtain robust estimates for
a confidence interval of the mean if the sample size is small (e.g. n = 10)
to directly calculate them from a theoretical normal distribution. Note
that the usual calculation based on quantiles of the theoretical
distribution can also be obtained with this function using the default
ci_type = "norm". This function provides a simplified user interface to the
\code{\link[boot]{boot}} and \code{\link[boot]{boot.ci}} functions
similarly to the \code{\link[simpleboot]{one.boot}} function but retains
more of the boot package's functionality, most notably including options
for parallelization. For convenience, when operating in parallel the user's
operating system is automatically detected so that the appropriate
parallelization engine is used (e.g. snow for Windows, multicore otherwise)
by the parallel package. Since the mean and median are common descriptive
statistics for which confidence intervals are estimated, these have their
own dedicated functions. To obtain bootstrapped confidence intervals for
other summary statistics use code{\link{stat_ci}} instead.
}
\examples{

y1 <- rnorm(1:10000, 100, 10)

#using a single core (sequential processing)
mean_ci(y1)

mean_ci(y1, ci_type = "perc")

#using multiple cores (parallel processing)
mean_ci(y1, parallel = TRUE, cores = 2, ci_type = "perc")

}
\references{
Efron, B., & Tibshirani, R. J. (1993). An introduction to the bootstrap. New
York: Chapman & Hall.
}
\seealso{
\code{\link[boot]{boot}}, \code{\link[boot]{boot.ci}},
\code{\link{median_ci}}, \code{\link{stat_ci}}, \code{\link{describe_ci}},
\code{\link{describe_ci_all}}
}
\author{
Craig P. Hutton, \email{craig.hutton@gmail.com}
}
