% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{mode}
\alias{mode}
\title{Obtain the mode of a vector.}
\usage{
mode(y, digits = 3, inv = FALSE, na.rm = TRUE)
}
\arguments{
\item{y}{a numeric vector/variable.}

\item{digits}{This determines the number of digits used for rounding of
numeric outputs. Default = 3.}

\item{inv}{This allows you to get the inverse or opposite of the mode, i.e.
the least common value or "anti-mode". Default is FALSE.}

\item{na.rm}{This determines whether missing values (NAs) should be removed
before attempting to count values and extract the mode (or anti-mode).}
}
\description{
\code{mode} returns the mode AKA most common value rounded to 3 decimal
places if y is a numeric vector. If y is non-numeric the most common value is
returned as a character string. For additional unique values of y sorted by
decreasing frequency, use counts() instead. N.B. this function overwrites a
base R function mode() with the same name that is a convenience shortcut used
to specify the storage mode of an object. However, this conflict isn't much
of an issue because that alternative function can still be accessed using the
full function name = "storage.mode()".
}
\examples{
y <- c(1:100, 2, 2, 4, 5, 6, 25, 50)

mode(y) #returns the mode

mode(y, inv = TRUE) #returns the anti-mode
}
\seealso{
\code{\link{counts}}
}
\author{
Craig P. Hutton, \email{Craig.Hutton@bov.bc.ca}
}
