% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{mode_of_y}
\alias{mode_of_y}
\title{Obtain the mode of a numeric variable.}
\usage{
mode_of_y(y)
}
\arguments{
\item{y}{a numeric vector/variable.}
}
\description{
\code{mode_of_y} returns the mode AKA most common value rounded to 3 decimal
places. For additional unique values of y sorted by decreasing frequency, use
mcvals instead
}
\examples{
y <- c(1:100)
mode_of_y(y)

}
\author{
Craig P. Hutton, \email{Craig.Hutton@bov.bc.ca}
}
