% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Generate a histogram.}
\usage{
plot_histogram(
  data,
  x,
  ...,
  binwidth = 1,
  bins = NULL,
  position = c("identity", "stack", "dodge"),
  stat = c("bin", "count"),
  na.rm = TRUE,
  fill_var = NULL,
  colour_var = NULL,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  fill_var_title = NULL,
  colour_var_title = NULL,
  xlim = c(NA, NA),
  transform_x = FALSE,
  x_transformation = "log10",
  fill_var_order = NULL,
  colour_var_order = NULL,
  fill_var_labs = NULL,
  colour_var_labs = NULL,
  fill_var_values = NULL,
  colour_var_values = NULL,
  alpha = 0.6,
  greyscale = FALSE,
  theme = "classic",
  text_size = 14,
  font = c("sans", "serif", "mono"),
  facet_var = NULL,
  facet_var_order = NULL,
  facet_var_labs = NULL,
  facet_var_strip_position = c("top", "bottom"),
  facet_var_text_bold = TRUE,
  legend_position = c("right", "left", "top", "bottom"),
  omit_legend = FALSE,
  interactive = FALSE,
  aesthetic_options = FALSE
)
}
\arguments{
\item{data}{A data frame or tibble containing the dependent measure "x" and
any grouping variables.}

\item{x}{The numeric variable you want a histogram of.}

\item{...}{graphical parameters (not associated with variables) to be passed
to \code{\link[ggplot2]{geom_histogram}}, e.g. colour or fill, to be applied
to all bars. To see some of the available options in a web browser, set
the aesthetic_options argument to TRUE.}

\item{binwidth}{Determines the aggregation level for bin construction in
units of x. Default = 1.}

\item{bins}{Instead of specifying the binsize using binwidth, you can specify
a total number of bins to display. If you choose to do so, set binwidth =
NULL.}

\item{position}{Determines how bars are organized when a grouping variable is
assigned to fill or colour. Valid options include "identity" (the
default),"stack", and "dodge"}

\item{stat}{Determines how the y-axis is constructed. Typically one would not
change this from the default of "bin" for a histogram, although "count" is
another sensible option. See \code{\link[ggplot2]{stat_bin}} &
\code{\link[ggplot2]{stat_count}} for details.}

\item{na.rm}{If set to TRUE (the default), missing values will be omitted
before plotting.}

\item{fill_var}{Use if you want to assign a variable to the histogram bar
fill colour, e.g. fill_var = grouping_variable. Produces separate sets of
bars for each level of the fill variable. See \code{\link[ggplot2]{aes}}
for details.}

\item{colour_var}{Use if you want to assign a variable to the histogram bar
outline colour, e.g. colour_var = grouping_variable. Produces separate sets
of bars for each level of the colour variable. See
\code{\link[ggplot2]{aes}} for details.}

\item{xlab}{Specify/overwrite the x-axis label using a character string, e.g.
"x-axis label"}

\item{ylab}{Specify/overwrite the y-axis label using a character string, e.g.
"y-axis label"}

\item{title}{Add a main title to the plot using a character string, e.g.
"Distribution of X"}

\item{fill_var_title}{If a variable has been assigned to fill using fill_var,
this allows you to modify the variable label in the plot legend.}

\item{colour_var_title}{If a variable has been assigned to colour using
colour_var, this allows you to modify the variable label in the plot
legend.}

\item{xlim}{specify the x-axis limits, e.g. xlim = c(lower_limit,
upper_limit). Use NA for the existing minimum or maximum value of x, e.g.
the default is xlim = c(NA, NA)}

\item{transform_x}{Would you like to transform the x axis? (TRUE or FALSE)}

\item{x_transformation}{If transform_x = TRUE, this determines the
transformation to be applied. Common choices include "log10" (the default),
"log2", "sqrt", or "exp". See \code{\link[ggplot2]{scale_continuous}} for
details.}

\item{fill_var_order}{If a variable has been assigned to fill using fill_var,
this allows you to modify the order of the variable groups, e.g. fill_var =
grouping_variable, fill_var_order = c("group_2", "group_1"). See
\code{\link[forcats]{fct_relevel}} for details.}

\item{colour_var_order}{If a variable has been assigned to colour using
colour_var, this allows you to modify the order of the variable groups,
e.g. colour_var = grouping_variable, fill_var_order = c("group_2",
"group_1"). See \code{\link[forcats]{fct_relevel}} for details.}

\item{fill_var_labs}{If a variable has been assigned to fill using fill_var,
this allows you to modify the labels of the variable groups, e.g. fill_var
= grouping_variable, fill_var_labs = c("group_1_new_label" =
"group_1_old_label", "group_2_new_label" = "group_2_old_label"). See
\code{\link[forcats]{fct_recode}} for details.}

\item{colour_var_labs}{If a variable has been assigned to colour using
colour_var, this allows you to modify the labels of the variable groups,
e.g. colour_var = grouping_variable, colour_var_labs =
c("group_1_new_label" = "group_1_old_label", "group_2_new_label" =
"group_2_old_label"). See \code{\link[forcats]{fct_recode}} for details.}

\item{fill_var_values}{If a variable has been assigned to fill using
fill_var, this allows you to modify the colours assigned to the fill of
each of the variable groups, e.g. fill_var = grouping_variable,
fill_var_values = c("blue", "red"). See
\code{\link[ggplot2]{scale_fill_manual}} for details. For the colour
options available in base R, see \code{\link[elucidate]{colour_options}}.}

\item{colour_var_values}{If a variable has been assigned to colour using
colour_var, this allows you to modify the colours assigned to the outline
of each of the variable groups, e.g. colour_var = grouping_variable,
colour_var_values = c("blue", "red"). See
\code{\link[ggplot2]{scale_fill_manual}} for details. For the colour
options available in base R, see \code{\link[elucidate]{colour_options}}.}

\item{alpha}{This adjusts the transparency/opacity of the graphical
components of the plot, ranging from 0 = 100\% transparent to 1 = 100\%
opaque.}

\item{greyscale}{Set to TRUE if you want the plot converted to greyscale.}

\item{theme}{Adjusts the theme using 1 of 6 predefined "complete" theme
templates provided by ggplot2. Currenlty supported options are: "classic"
(the elucidate default), "bw", "grey" (the ggplot2 default), "light",
"dark", & "minimal". See \code{\link[ggplot2]{theme_classic}} for more
information.}

\item{text_size}{This controls the size of all plot text. Default = 14.}

\item{font}{This controls the font of all plot text. Default = "sans"
(Arial). Other options include "serif" (Times New Roman) and "mono" (Courier
New).}

\item{facet_var}{Use if you want separate plots for each level of a grouping
variable (i.e. a facetted plot), e.g. facet_var = grouping_variable. See
\code{\link[ggplot2]{facet_wrap}} for details.}

\item{facet_var_order}{If a variable has been assigned for facetting using
facet_var, this allows you to modify the order of the variable groups, e.g.
facet_var = grouping_variable, facet_var_order = c("group_2", "group_1").
See \code{\link[forcats]{fct_relevel}} for details.}

\item{facet_var_labs}{If a variable has been assigned for facetting using
facet_var, this allows you to modify the labels of the variable groups
which will appear in the facet strips, e.g. facet_var = grouping_variable,
facet_var_labs = c("group_1_new_label" = "group_1_old_label",
"group_2_new_label" = "group_2_old_label"). See
\code{\link[forcats]{fct_recode}} for details.}

\item{facet_var_strip_position}{If a variable has been assigned for facetting
using facet_var, this allows you to modify the position of the facet strip
labels. Sensible options include "top" (the default) or "bottom".}

\item{facet_var_text_bold}{If a variable has been assigned for facetting
using facet_var, this allows you to use boldface (TRUE/default or FALSE)
for the facet strip label text.}

\item{legend_position}{This allows you to modify the legend position.
Options include "right" (the default), "left", "top", & "bottom".}

\item{omit_legend}{Set to TRUE if you want to remove/omit the legends.}

\item{interactive}{Determines whether a static ggplot object or an interactive html
plotly object is returned. See \code{\link[plotly]{ggplotly}} for details.}

\item{aesthetic_options}{If set to TRUE, opens a web browser to the tidyverse
online aesthetic options vignette.}
}
\value{
A ggplot object or plotly object depending on whether static or
interactive output was requested.
}
\description{
Easily generate a histogram of a variable using
ggplot2 with a simplified customization interface for common modifications
with static (ggplot) and interactive (plotly) output options. The static
output is useful for producing static reports (e.g. for manuscripts) and is
readily customized further using ggplot2 syntax. The interactive output is
helpful for exploring the data and producing dynamic html reports.
}
\examples{
data(mtcars) #load the mtcars data

library(magrittr)

mtcars \%>\% plot_histogram(x = mpg, fill = "blue")

\dontrun{
mtcars \%>\% plot_histogram(x = mpg, fill = "blue")

mtcars \%>\% plot_histogram(x = mpg, fill = "blue", colour = "black")

mtcars \%>\% plot_histogram(x = mpg, colour_var = cyl, fill = "white")

mtcars \%>\% plot_histogram(x = mpg, fill_var = cyl, position = "identity") #default in elucidate

mtcars \%>\% plot_histogram(x = mpg, fill_var = cyl, position = "dodge", binwidth = 1)

mtcars \%>\% plot_histogram(x = mpg, fill_var = cyl, position = "stack") #default in ggplot2

mtcars \%>\% plot_histogram(x = mpg, fill = "blue", binwidth = 5)

mtcars \%>\% plot_histogram(x = mpg, fill = "blue", binwidth = NULL, bins = 30) #default in ggplot2

mtcars \%>\% plot_histogram(x = mpg, fill = "blue", interactive = T)

mtcars \%>\% plot_histogram(x = mpg, fill_var = cyl, binwidth = 5, interactive = T)
}


}
\references{
Wickham, H. (2016). ggplot2: elegant graphics for data analysis. New York, N.Y.: Springer-Verlag.
}
\seealso{
\code{\link[ggplot2]{geom_histogram}}, \code{\link[plotly]{ggplotly}}
}
\author{
Craig P. Hutton, \email{Craig.Hutton@gov.bc.ca}
}
