% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_var_all}
\alias{plot_var_all}
\title{Plot all variables/vectors in a data frame using a class-appropriate \verb{plot_*}
function.}
\usage{
plot_var_all(
  data,
  var2 = NULL,
  group_var = NULL,
  cols = NULL,
  var2_lab = ggplot2::waiver(),
  title = ggplot2::waiver(),
  fill = "blue2",
  colour = "black",
  palette = c("plasma", "C", "magma", "A", "inferno", "B", "viridis", "D", "cividis",
    "E"),
  palette_direction = c("d2l", "l2d"),
  palette_begin = 0,
  palette_end = 0.8,
  alpha = 0.75,
  greyscale = FALSE,
  line_size = 1,
  theme = c("classic", "bw", "grey", "light", "dark", "minimal"),
  text_size = 14,
  font = c("sans", "serif", "mono"),
  legend_position = c("right", "left", "top", "bottom"),
  omit_legend = FALSE,
  dnorm = TRUE,
  violin = TRUE,
  var1_log10 = FALSE,
  var2_log10 = FALSE,
  point_size = 2,
  point_shape = c("circle", "square", "diamond", "triangle up", "triangle down"),
  regression_line = TRUE,
  regression_method = c("gam", "loess", "lm"),
  regression_se = TRUE,
  bar_position = c("dodge", "fill", "stack"),
  bar_width = 0.9,
  basic = FALSE,
  interactive = FALSE,
  trelliscope = FALSE,
  nrow = NULL,
  ncol = NULL,
  guides = "collect"
)
}
\arguments{
\item{data}{A data frame containing variables to be plotted.}

\item{var2}{The (quoted or unquoted) name of a secondary variable to plot
against all other variables in the input data (or a subset of them if the
\code{cols} argument is used), where the latter set of "primary" variables will
be automatically assigned to the \code{var1} argument of \code{\link{plot_var}}.
\code{var2} is usually assigned to the x-axis. However, if the primary variable
(i.e. \code{var1}) is a categorical (factor, character, or logical) variable and
\code{var2} is a numeric, integer, or date variable, \code{var2} will be assigned to
the y-axis and \code{var1} will be assigned to the x-axis. If \code{var1} and \code{var2}
are both categorical variables, \code{var1} will be assigned to the x-axis and
\code{var2} will be assigned to \code{facet_var}.}

\item{group_var}{Use if you want to assign a grouping variable to fill
(colour) and/or (outline) colour e.g. group_var = "grouping_variable" or
group_var = grouping_variable. Whether the grouping variable is mapped to
fill, colour, or both will depend upon which \verb{plot_*} function is used (See
"Value" section). For density plots, both fill and colour are used for
consistency across the main density plots and added normal density curve
lines (if dnorm = TRUE). For bar graphs and box-and-whisker plots, the
variable will be assigned to fill. For scatter plots, the variable will be
assigned to colour. See \code{\link[ggplot2]{aes}} for details.}

\item{cols}{A character (or integer) vector of column names (or indices)
which allows you to plot only a subset of the columns in the input data
frame, where each of these primary variable columns will be automatically
assigned to the \code{var1} argument of \code{\link{plot_var}}. Note that a
variable which has been assigned to \code{var2} or \code{group_var} does not also
need to be listed here.}

\item{var2_lab}{Accepts a character string to use to change the axis label
for the variable assigned to \code{var2}. Ignored if \code{var2} and the primary
variable are both categorical variables (since \code{var2} will be used for
faceting in such cases).}

\item{title}{A character string to add as a title at the top of the combined
multiple-panel patchwork graph or trelliscopejs display.}

\item{fill}{Fill colour to use for density plots, bar graphs, and box plots.
Ignored if a variable that has been assigned to \code{group_var} is mapped on to
\code{fill_var} (see \code{group_var} argument information above). Default is
"blue2". Use \code{\link{colour_options}} to see colour option examples.}

\item{colour}{Outline colour to use for density plots, bar graphs, box plots,
and scatter plots. Ignored if a variable that has been assigned to
\code{group_var} is mapped on to \code{colour_var} (see \code{group_var} argument
information above). Default is "black". Use \code{\link{colour_options}} to
see colour option examples.}

\item{palette}{If a variable is assigned to group_var, this determines which
\href{https://cran.r-project.org/web/packages/viridis/vignettes/intro-to-viridis.html}{viridis
colour palette} to use. Options include "plasma" or "C" (default), "magma"
or "A", "inferno" or "B", "viridis" or "D", and "cividis" or "E". See
\href{https://craig.rbind.io/post/2021-05-17-asgr-3-1-data-visualization/#colourblind-friendly-palettes}{this
link} for examples.}

\item{palette_direction}{Choose "d2l" for dark to light (default) or "l2d"
for light to dark.}

\item{palette_begin}{Value between 0 and 1 that determines where along the
full range of the chosen colour palette's spectrum to begin sampling
colours. See \code{\link[ggplot2]{scale_fill_viridis_d}} for details.}

\item{palette_end}{Value between 0 and 1 that determines where along the full
range of the chosen colour palette's spectrum to end sampling colours. See
\code{\link[ggplot2]{scale_fill_viridis_d}} for details.}

\item{alpha}{This adjusts the transparency/opacity of the main geometric
objects in the generated plot, with acceptable values ranging from 0 = 100\%
transparent to 1 = 100\% opaque.}

\item{greyscale}{Set to TRUE if you want the plot converted to grey scale.}

\item{line_size}{Controls the thickness of plotted lines.}

\item{theme}{Adjusts the theme using 1 of 6 predefined "complete" theme
templates provided by ggplot2. Currently supported options are: "classic"
(the elucidate default), "bw", "grey" (the ggplot2 default), "light",
"dark", & "minimal". See \code{\link[ggplot2]{theme_classic}} for more
information.}

\item{text_size}{This controls the size of all plot text. Default = 14.}

\item{font}{This controls the font of all plot text. Default = "sans"
(Arial). Other options include "serif" (Times New Roman) and "mono" (Courier
New).}

\item{legend_position}{This allows you to modify the legend position if a
variable is assigned to \code{group_var}. Options include "right" (the default),
"left", "top", & "bottom".}

\item{omit_legend}{Set to TRUE if you want to remove/omit the legend(s).
Ignored if \code{group_var} is unspecified.}

\item{dnorm}{When TRUE (default), this adds a dashed line representing a
normal/Gaussian density curve to density plots, which are rendered for
plots of single numeric variables. Disabled if \code{var1} is a date vector,
\code{var1_log10} = TRUE, or \code{basic} = TRUE.}

\item{violin}{When TRUE (default), this adds violin plot outlines to box
plots, which are rendered in cases where a mixture of numeric and
categorical variables are assigned to \code{var1} and \code{var2}. Disabled if
\code{basic} = TRUE.}

\item{var1_log10}{If TRUE, applies a base-10 logarithmic transformation to a
numeric variable that has been assigned to \code{var1}. Ignored if \code{var1} is a
categorical variable.}

\item{var2_log10}{If TRUE, applies a base-10 logarithmic transformation to a
numeric variable that has been assigned to \code{var2}. Ignored if \code{var2} is a
categorical variable.}

\item{point_size}{Controls the size of points used in scatter plots, which
are rendered in cases where \code{var1} and \code{var2} are both numeric, integer, or
date variables.}

\item{point_shape}{Point shape to use in scatter plots, which
are rendered in cases where \code{var1} and \code{var2} are both numeric, integer, or
date variables.}

\item{regression_line}{If TRUE (the default), adds a regression line to scatter
plots, which are rendered in cases where \code{var1} and \code{var2} are both
numeric, integer, or date variables. Disabled if \code{basic} = TRUE.}

\item{regression_method}{If \code{regression_line} = TRUE, this determines the
type of regression line to use. Currently available options are "gam",
"loess", and "lm". "gam" is the default, which fits a generalized additive
model using a smoothing term for x. This method has a longer run time, but
typically provides a better fit to the data than other options and uses an
optimization algorithm to determine the optimal wiggliness of the line. If
the relationship between y and x is linear, the output will be equivalent
to fitting a linear model. "loess" may be preferable to "gam" for small
sample sizes. See \code{\link[ggplot2]{stat_smooth}} and
\code{\link[mgcv]{gam}} for details.}

\item{regression_se}{If TRUE (the default), adds a 95\% confidence envelope for the
regression line. Ignored if \code{regression_line} = FALSE.}

\item{bar_position}{In bar plots, which are rendered for one or more
categorical variables, this determines how bars are arranged relative to
one another when a grouping variable is assigned to \code{group_var}. The
default, "dodge", uses \code{\link[ggplot2]{position_dodge}} to arrange
bars side-by-side; "stack" places the bars on top of each other; "fill"
also stacks bars but additionally converts y-axis from counts to
proportions.}

\item{bar_width}{In bar plots, which are rendered for one or more categorical
variables, this adjusts the width of the bars (default = 0.9).}

\item{basic}{This is a shortcut argument that allows you to simultaneously
disable the \code{dnorm}, \code{violin}, and \code{regression_line} arguments to produce a
basic version of a density, box, or scatter plot (depending on
\code{var1}/\code{var2} variable class(es)) without any of those additional layers.
Dropping these extra layers may noticeably reduce rendering time and memory
utilization, especially for larger sample sizes and/or when \code{interactive} =
TRUE.}

\item{interactive}{Determines whether a static ggplot object or an
interactive html plotly object is returned. Interactive/plotly mode for
multiple plots should only be used in conjunction with \code{trelliscope} =
TRUE. See \code{\link[plotly]{ggplotly}} for details. Note that in cases
where a box plot is generated (for a mix of numeric and categorical
variables) and a variable is also assigned to \code{group_var}, activating
interactive/plotly mode will cause a spurious warning message about
'layout' objects not having a 'boxmode' attribute to be printed to the
console. This is a
\href{https://github.com/ropensci/plotly/issues/994}{documented bug} with
plotly that can be safely ignored, although unfortunately the message
cannot currently be suppressed.}

\item{trelliscope}{If changed to TRUE, plots will be combined into an
interactive trelliscope display rather than a static patchwork graph grid.
See \code{\link[trelliscopejs]{trelliscope}} for more information.}

\item{nrow}{This controls the number of rows to use when arranging plots in
the combined patchwork or trelliscopejs display.}

\item{ncol}{This controls the number of columns to use when arranging plots
in the combined patchwork or trelliscopejs display.}

\item{guides}{Controls the pooling of \code{group_var} legends/guides across plot
panels if a categorical variable has been assigned to \code{group_var} and
\code{trelliscope} = FALSE. See \code{\link[patchwork]{wrap_plots}} for details.}
}
\value{
A static "patchwork" or dynamic "trelliscope" multi-panel graphical
display of ggplot2 or plotly graphs depending upon the values of the
\code{trelliscope} and \code{interactive} arguments. The type of graph (i.e.
\verb{ggplot2::geom*} layers) that is rendered in each panel will depend upon
the classes of the chosen variables, as follows:
\itemize{
\item One numeric (classes numeric/integer/date) variable will be graphed with
\code{\link{plot_density}}.
\item One or two categorical (classes factor/character/logical) variable(s)
will be graphed with \code{\link{plot_bar}}.
\item Two numeric variables will be graphed with \code{\link{plot_scatter}}.
\item A mixture of numeric and categorical variables will be graphed with
\code{\link{plot_box}}.
}
}
\description{
Easily generate ggplot2 graphs for all (or a named vector of)
variables in a data frame using a class-appropriate geometry via other
elucidate \verb{plot_*} functions with a restricted set of customization options
and some modified defaults. The \code{var2} argument also allows you to plot all
variables against a specific named secondary variable. The collection of
generated graphs will be combined into a single lattice-style figure with
either the patchwork package or trelliscopejs package. See "Arguments"
section for details and
\href{https://craig.rbind.io/post/2021-05-17-asgr-3-1-data-visualization/}{this
blog post} for an introduction to ggplot2. To obtain a plot of a single
variable or vector, use \code{\link{plot_var}} instead. To obtain pairwise plots of
all bivariate combinations of variables, use \code{\link{plot_var_pairs}}
instead.
}
\examples{

data(mtcars) #load the mtcars data

#convert variables "cyl" to a factors
mtcars$cyl <- as.factor(mtcars$cyl)

#plot variables "hp", "wt", and "cyl" from the mtcars data frame
plot_var_all(mtcars, cols = c("hp", "wt", "cyl"))

#plot each of the same variables against column "mpg"
plot_var_all(mtcars, var2 = mpg, cols = c("hp", "wt", "cyl"))

#plot "hp" and "wt" against mpg, group by "cyl"
plot_var_all(mtcars, var2 = mpg, group_var = cyl, cols = c("hp", "wt"),
             basic = TRUE, #distable regression lines/CIs
             ncol = 1, nrow = 2) #change the layout

}
\references{
Wickham, H. (2016). ggplot2: elegant graphics for data analysis. New York, N.Y.: Springer-Verlag.
}
\seealso{
\code{\link{plot_density}}, \code{\link{plot_bar}},
\code{\link{plot_scatter}}, \code{\link{plot_box}}, \code{\link{plot_var}},
\code{\link{plot_var_pairs}}, \code{\link[patchwork]{wrap_plots}},
\code{\link[trelliscopejs]{trelliscope}}
}
\author{
Craig P. Hutton, \email{Craig.Hutton@gov.bc.ca}
}
