% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{se}
\alias{se}
\title{Calculate the standard error of the mean.}
\usage{
se(y)
}
\arguments{
\item{y}{a numeric vector/variable.}
}
\description{
\code{se} returns the standard error of the mean.
}
\examples{
y <- c(1:100)
se(y)

}
\references{
Altman, D. G., & Bland, J. M. (2005). Standard deviations and standard
errors. Bmj, 331(7521), 903.
}
\seealso{
\code{\link[stats]{sd}}, \code{\link{sqrt}}, \code{\link{length}}
}
\author{
Craig P. Hutton, \email{craig.hutton@gmail.com}
}
