% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{skewness}
\alias{skewness}
\title{Calculate the skewness of a numeric variable.}
\usage{
skewness(y, na.rm = TRUE, type = 2)
}
\arguments{
\item{y}{a numeric vector/variable.}

\item{na.rm}{should missing values be removed before attempting to calculate
skewness? Default is TRUE.}

\item{type}{the calculation method you wish to use. See
\code{\link[e1071]{skewness}} and Joanes & Gill (1998) for details.}
}
\description{
\code{skewness} calculates the degree of skewness in a numerical
distribution. Implements the same functionality as the skewness functions in
the psych and e1071 packages for convenience so that those packages are not
required as dependencies for elucidate. Unlike the e1071 and psych packages,
the elucidate version uses the type 2 calculation as a default since this
incorporates a correction for sample size that also matches what is used in
SPSS and SAS.
}
\examples{
y <- rnorm(1:1000, 100, 15)
skewness(y)

}
\references{
D. N. Joanes and C. A. Gill (1998), Comparing measures of sample skewness and
kurtosis. The Statistician, 47, 183–189.
}
\seealso{
\code{\link[e1071]{skewness}}, \code{\link[psych]{skew}}
}
\author{
Craig P. Hutton, \email{Craig.Hutton@bov.bc.ca}
}
