% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{static_to_dynamic}
\alias{static_to_dynamic}
\title{Convert a static data frame or ggplot object to a dynamic form.}
\usage{
static_to_dynamic(static_object, caption = NULL)
}
\arguments{
\item{static_object}{A data frame, tibble, or ggplot2 object.}

\item{caption}{Add a caption/title to the dynamic table/figure.}
}
\value{
If a data frame or tibble was provided, the output will be a
datatables html widget. If a ggplot object was provided, the output will be
a plotly html widget.
}
\description{
Uses \code{\link[plotly]{ggplotly}} or
\code{\link[DT]{datatable}} to convert a ggplot2 object or data
frame/tibble to a dynamic & interactive form.
}
\examples{

\dontrun{
data(mtcars)
library(ggplot2)

p1 <- ggplot(mtcars, aes(x = mpg)) + geom_histogram(binwidth = 1)
p1 \%>\% static_to_dynamic(caption = "Figure 1")

mtcars \%>\% static_to_dynamic(caption = "Table 1")
}

}
\seealso{
\code{\link[DT]{datatable}}, \code{\link[plotly]{ggplotly}}
}
\author{
Craig P. Hutton, \email{Craig.Hutton@bov.bc.ca}
}
