% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{tcv}
\alias{tcv}
\title{elucidate package internal function}
\usage{
tcv(y, n = "all", order = "d")
}
\arguments{
\item{y}{A vector/variable (required).}

\item{n}{The number of unique values you want frequency counts for. Default is "all".}

\item{order}{"d" for descending/decreasing order. "a" or "i" for
ascending/increasing order.}
}
\description{
\code{tcv} is an internal function that supports
\code{\link{describe}}.
}
\author{
Craig P. Hutton, \email{craig.hutton@gmail.com}
}
