% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Obtain a descriptive summary of a variable.}
\usage{
describe(data, y = NULL, ..., digits = 3, order = "d", n = "all",
  type = 2, na.rm = TRUE, output = "tibble")
}
\arguments{
\item{data}{Either a vector or a data frame or tibble containing the
vector ("y") to be summarized and any grouping variables.}

\item{y}{If the data object is a data.frame, this is the variable for which
you wish to obtain a descriptive summary}

\item{...}{If the data object is a data.frame, this special argument accepts
any number of unquoted grouping variable names (also present in the data source)
to use for subsetting, separated by commas (e.g. \code{group_var1,
  group_var2})}

\item{digits}{This determines the number of digits used for rounding of
numeric outputs.}

\item{order}{For variables/vectors of class "character" or "factor" this
determines if the unique values should be sorted in order of
descending/decreasing = "d" or ascending/increasing = "a" or "i" frequency
counts.}

\item{n}{For variables/vectors of class "character" or "factor" this
determines how many of the top unique values based on increasing or
decreasing frequency (as specified by the order argument) are
saved/printed. E.g. for just the most common value, set order = "d"
(default) and n = 1.}

\item{type}{For numeric and integer vectors this determines the type of
skewness and kurtosis calculations to perform. See
\code{\link[e1071]{skewness}} or \code{\link[psych]{skew}} and
\code{\link[e1071]{kurtosis}} or \code{\link[psych]{kurtosi}} for details.}

\item{na.rm}{This determines whether missing values (NAs) should be removed
before attempting to calculate summary statistics.}

\item{output}{"tibble" for tibble or "dt" for data.table. Tibble is used as
the default output to facilitate subsequent use/modification of the output
with the tidyverse collection of packages.}
}
\value{
The output varies as a function of the class of input data/y, referred to as "y" below

\strong{For all input variables, the following are returned (part 1):}

\describe{
\item{cases}{the total number of cases}
\item{n}{number of complete cases}
\item{na}{the number of missing values}
\item{cases}{the proportion of total cases with missing values}
}

In addition to part 1, these measures are provided for \strong{dates}:

\describe{
\item{n_unique}{the total number of unique values or levels of y. For dates this tells you how many time points there are}
\item{start}{the earliest or minimum date in y}
\item{end}{the latest or maximum date in y}
\item{p_FALSE}{the proportion of y values that are FALSE}
}

In addition to part 1, these measures are provided for \strong{factors}:

\describe{
\item{n_unique}{the total number of unique values or levels of y}
\item{ordered}{a logical indicating whether or not y is ordinal}
\item{value_n}{a series of outputs for the top "n" or "all" unique values of y with their frequency counts following the format "value_count"}
}

In addition to part 1, these measures are provided for \strong{character/string} vectors:

\describe{
\item{n_unique}{the total number of unique values or levels of y}
\item{value_n}{a series of outputs for the top "n" or "all" unique values of y with their frequency counts following the format "value_count"}
}

In addition to part 1, these measures are provided for \strong{logcial} vectors:

\describe{
\item{n_TRUE}{the total number of y values that are TRUE}
\item{n_FALSE}{the total number of y values that are FALSE}
\item{p_TRUE}{the proportion of y values that are TRUE}
\item{p_FALSE}{the proportion of y values that are FALSE}
}

In addition to part 1, these measures are provided for \strong{numeric} variables:

\describe{
\item{mean}{the mean of y}
\item{sd}{the standard deviation of y}
\item{se}{the standard error of the mean of y}
\item{p0}{the 0th percentile (the minimum) of y}
\item{p25}{the 25th percentile of y}
\item{p50}{the 50th percentile (the median) of y}
\item{p75}{the 25th percentile of y}
\item{p100}{the 100th percentile (the maximum) of y}
\item{skew}{the skewness of the distibution of y}
\item{kurt}{the kurtosis of the distibution of y}
}
}
\description{
Obtain a useful array of common summary statistics for a
vector/variable with customized output depending on the class of variable.
Uses a combination of tidyverse packages and data.table to provide a
user-friendly interface that is pipe-friendly while leveraging the
excellent performance of data.table. The use of the ... argument also makes
it incredibly easy to obtain summaries split by grouping variables. While
other similar functions exist in other packages (e.g.
\code{\link[psych]{describeBy}} or \code{\link[skimr]{skim}}), this version
provides the some of the useful added outputs of the psych package (e.g.
skew and kurtosis for numeric variables) while at the same time offering
slightly more concise syntax than skim (e.g. no preceding group_by
operation is needed for group-wise calculations) and enhanced customization
options via clearly defined arguments than the alternatives (e.g. ascending
or descending sorting, modifying the number of unique values to return
counts for character/numeric vectors) while still achieving comparable
processing times to the alternatives. To obtain summaries for all variables
in a data frame use \code{\link{describe_all}} instead.
}
\examples{

describe(data = pdata, y = y1) #no grouping variables, numeric input class
describe(pdata, y1, high_low) #one grouping variable, numeric input class
describe(pdata, g2) #factor input class
describe(pdata, even) #logical input class

}
\references{
Altman, D. G., & Bland, J. M. (2005). Standard deviations and standard
errors. Bmj, 331(7521), 903.

Bulmer, M. G. (1979). Principles of statistics. Courier Corporation.

D. N. Joanes and C. A. Gill (1998), Comparing measures of sample skewness and
kurtosis. The Statistician, 47, 183-189.
}
\seealso{
\code{\link[base]{mean}}, \code{\link[stats]{sd}}, \code{\link{se}},
\code{\link[stats]{quantile}}, \code{\link{skewness}}, \code{\link{kurtosis}}
}
\author{
Craig P. Hutton, \email{craig.hutton@gmail.com}
}
