% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe_all}
\alias{describe_all}
\title{Obtain descriptive summaries for all variables in a data frame.}
\usage{
describe_all(data, ..., class = "all", digits = 3, order = "d",
  n = 5, type = 2, na.rm = TRUE, output = "tibble")
}
\arguments{
\item{data}{A data frame or tibble.}

\item{...}{This special argument accepts any number of unquoted grouping
variable names (also present in the data source) to use for subsetting,
separated by commas (e.g. \code{group_var1, group_var2})}

\item{class}{The variable classes in data that you would like summaries for.
Either "all" for all classes, or a character vector indicating which
combinations of output classes you want. Specifying a subset will save time
since summaries are only processed as needed. Options include "d" for
dates, "f" for factors, "c" for character, "l" for logical, and "n" for
numeric. If only a single class is requested or present in the data after
excluding specified grouping variables, a data frame will be returned,
otherwise you'll get a list of data frames (1 per summary class). If the
only chosen class of variables is not detected in the input data an error
will be returned that the class argument needs to be respecified.}

\item{digits}{This determines the number of digits used for rounding of
numeric outputs.}

\item{order}{For variables/vectors of class "character" or "factor" this
determines if the unique values should be sorted in order of
descending/decreasing = "d" or ascending/increasing = "a" or "i" frequency
counts.}

\item{n}{For variables/vectors of class "character" or "factor" this
determines how many of the top unique values based on increasing or
decreasing frequency (as specified by the order argument) are
saved/printed. E.g. for just the most common value, set order = "d"
(default) and n = 1.}

\item{type}{For numeric and integer vectors this determines the type of
skewness and kurtosis calculations to perform. See
\code{\link[e1071]{skewness}} or \code{\link[psych]{skew}} and
\code{\link[e1071]{kurtosis}} or \code{\link[psych]{kurtosi}} for details.}

\item{na.rm}{This determines whether missing values (NAs) should be removed
before attempting to calculate summary statistics.}

\item{output}{Output type for each class of variables. "tibble" for tibble or
"dt" for data.table. Tibble is used as the default output to facilitate
subsequent use/modification of the output with the tidyverse collection of
packages.}
}
\value{
The output varies as a function of the class of input data/y,
referred to as "y" below. Each output type is grouped together in a data
frame and returned as a named item of a list, unless there is only one
output type, in which case the data frame is returned directly.

\strong{For all input variables, the following are returned (part 1):}

\describe{
\item{cases}{the total number of cases}
\item{n}{number of complete cases}
\item{na}{the number of missing values}
\item{cases}{the proportion of total cases with missing values}
}

In addition to part 1, these measures are provided for \strong{dates}:

\describe{
\item{n_unique}{the total number of unique values or levels of y. For dates this tells you how many time points there are}
\item{start}{the earliest or minimum date in y}
\item{end}{the latest or maximum date in y}
\item{p_FALSE}{the proportion of y values that are FALSE}
}

In addition to part 1, these measures are provided for \strong{factors}:

\describe{
\item{n_unique}{the total number of unique values or levels of y}
\item{ordered}{a logical indicating whether or not y is ordinal}
\item{value_n}{a series of outputs for the top "n" or "all" unique values of y with their frequency counts following the format "value_count"}
}

In addition to part 1, these measures are provided for \strong{character/string} vectors:

\describe{
\item{n_unique}{the total number of unique values or levels of y}
\item{value_n}{a series of outputs for the top "n" or "all" unique values of y with their frequency counts following the format "value_count"}
}

In addition to part 1, these measures are provided for \strong{logcial} vectors:

\describe{
\item{n_TRUE}{the total number of y values that are TRUE}
\item{n_FALSE}{the total number of y values that are FALSE}
\item{p_TRUE}{the proportion of y values that are TRUE}
\item{p_FALSE}{the proportion of y values that are FALSE}
}

In addition to part 1, these measures are provided for \strong{numeric} variables:

\describe{
\item{mean}{the mean of y}
\item{sd}{the standard deviation of y}
\item{se}{the standard error of the mean of y}
\item{p0}{the 0th percentile (the minimum) of y}
\item{p25}{the 25th percentile of y}
\item{p50}{the 50th percentile (the median) of y}
\item{p75}{the 25th percentile of y}
\item{p100}{the 100th percentile (the maximum) of y}
\item{skew}{the skewness of the distibution of y}
\item{kurt}{the kurtosis of the distibution of y}
}
}
\description{
This function extends \code{{describe}} by applying to it all
columns of the specified class(es) in a data frame using functional
programming tools from the purrr package (e.g. \code{\link[purrr]{map}}).
To obtain a summary of a single variable in a data frame use
\code{\link{describe}} instead.
}
\examples{

describe_all(mtcars)

\dontrun{
describe_all(pdata) #all summary types in a list

#numeric summary only
pdata \%>\% describe_all(high_low, output = "dt", class = "n")

#numeric and logical summaries only
pdata \%>\%
describe_all(high_low, output = "dt", class = c("n", "l"))
}

}
\seealso{
\code{\link{describe}}
}
\author{
Craig P. Hutton, \email{craig.hutton@gmail.com}
}
