% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_density}
\alias{plot_density}
\title{Generate a density plot.}
\usage{
plot_density(
  data,
  x,
  ...,
  fill_var = NULL,
  colour_var = NULL,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  fill_var_title = NULL,
  colour_var_title = NULL,
  xlim = c(NA, NA),
  transform_x = FALSE,
  x_transformation = "log10",
  fill_var_order = NULL,
  colour_var_order = NULL,
  fill_var_labs = NULL,
  colour_var_labs = NULL,
  fill_var_values = NULL,
  colour_var_values = NULL,
  alpha = 0.6,
  greyscale = FALSE,
  theme = "classic",
  text_size = 14,
  font = "sans",
  font_options = FALSE,
  facet_var = NULL,
  facet_var_order = NULL,
  facet_var_labs = NULL,
  facet_var_strip_position = "top",
  facet_var_text_bold = TRUE,
  legend_position = "right",
  omit_legend = FALSE,
  interactive = FALSE,
  aesthetic_options = FALSE
)
}
\arguments{
\item{data}{A data frame or tibble containing the dependent measure "x" and
any grouping variables.}

\item{x}{The numeric variable you want a density plot of.}

\item{...}{graphical parameters (not associated with variables) to be passed
to \code{\link[ggplot2]{geom_density}}, e.g. colour or fill, to be applied
to all curves. To see some of the available options in a web browser, set
the aesthetic_options argument to TRUE.}

\item{fill_var}{Use if you want to assign a variable to the density curve
fill colour, e.g. fill_var = grouping_variable. Produces separate curves
for each level of the fill variable. See \code{\link[ggplot2]{aes}} for
details.}

\item{colour_var}{Use if you want to assign a variable to the density curve
outline colour, e.g. colour_var = grouping_variable. Produces separate
curves for each level of the colour variable. See
\code{\link[ggplot2]{aes}} for details.}

\item{xlab}{Specify/overwrite the x-axis label using a character string, e.g.
"x-axis label"}

\item{ylab}{Specify/overwrite the y-axis label using a character string, e.g.
"y-axis label"}

\item{title}{Add a main title to the plot using a character string, e.g.
"Density graph of X"}

\item{fill_var_title}{If a variable has been assigned to fill using fill_var,
this allows you to modify the variable label in the plot legend.}

\item{colour_var_title}{If a variable has been assigned to colour using
colour_var, this allows you to modify the variable label in the plot
legend.}

\item{xlim}{specify the x-axis limits, e.g. xlim = c(lower_limit,
upper_limit). Use NA for the existing minimum or maximum value of x, e.g.
the default is xlim = c(NA, NA)}

\item{transform_x}{Would you like to transform the x axis? (TRUE or FALSE)}

\item{x_transformation}{If transform_x = TRUE, this determines the
transformation to be applied. Common choices include "log10" (the default),
"log2", "sqrt", or "exp". See \code{\link[ggplot2]{scale_continuous}} for
details.}

\item{fill_var_order}{If a variable has been assigned to fill using fill_var,
this allows you to modify the order of the variable groups, e.g. fill_var =
grouping_variable, fill_var_order = c("group_2", "group_1"). See
\code{\link[forcats]{fct_relevel}} for details.}

\item{colour_var_order}{If a variable has been assigned to colour using
colour_var, this allows you to modify the order of the variable groups,
e.g. colour_var = grouping_variable, fill_var_order = c("group_2",
"group_1"). See \code{\link[forcats]{fct_relevel}} for details.}

\item{fill_var_labs}{If a variable has been assigned to fill using fill_var,
this allows you to modify the labels of the variable groups, e.g. fill_var
= grouping_variable, fill_var_labs = c("group_1_new_label" =
"group_1_old_label", "group_2_new_label" = "group_2_old_label"). See
\code{\link[forcats]{fct_recode}} for details.}

\item{colour_var_labs}{If a variable has been assigned to colour using
colour_var, this allows you to modify the labels of the variable groups,
e.g. colour_var = grouping_variable, colour_var_labs =
c("group_1_new_label" = "group_1_old_label", "group_2_new_label" =
"group_2_old_label"). See \code{\link[forcats]{fct_recode}} for details.}

\item{fill_var_values}{If a variable has been assigned to fill using
fill_var, this allows you to modify the colours assigned to the fill of
each of the variable groups, e.g. fill_var = grouping_variable,
fill_var_values = c("blue", "red"). See
\code{\link[ggplot2]{scale_fill_manual}} for details. For the colour
options available in base R, see \code{\link[elucidate]{colour_options}}.}

\item{colour_var_values}{If a variable has been assigned to colour using
colour_var, this allows you to modify the colours assigned to the outline
of each of the variable groups, e.g. colour_var = grouping_variable,
colour_var_values = c("blue", "red"). See
\code{\link[ggplot2]{scale_fill_manual}} for details. For the colour
options available in base R, see \code{\link[elucidate]{colour_options}}.}

\item{alpha}{This adjusts the transparency/opacity of the graphical
components of the plot, ranging from 0 = 100\% transparent to 1 = 100\%
opaque.}

\item{greyscale}{Set to TRUE if you want the plot converted to greyscale.}

\item{theme}{Adjusts the theme using 1 of 6 predefined "complete" theme
templates provided by ggplot2. Currenlty supported options are: "classic"
(the elucidate default), "bw", "grey" (the ggplot2 default), "light",
"dark", & "minimal". See \code{\link[ggplot2]{theme_classic}} for more
information.}

\item{text_size}{This controls the size of all plot text. Default = 14.}

\item{font}{This controls the font of all plot text. Default = "sans" (Arial).}

\item{font_options}{Set to TRUE if you want the (3) available font options to
be printed to the console. See \code{\link[grDevices]{windowsFonts}} for
details.}

\item{facet_var}{Use if you want separate plots for each level of a grouping
variable (i.e. a facetted plot), e.g. facet_var = grouping_variable. See
\code{\link[ggplot2]{facet_wrap}} for details.}

\item{facet_var_order}{If a variable has been assigned for facetting using
facet_var, this allows you to modify the order of the variable groups, e.g.
facet_var = grouping_variable, facet_var_order = c("group_2", "group_1").
See \code{\link[forcats]{fct_relevel}} for details.}

\item{facet_var_labs}{If a variable has been assigned for facetting using
facet_var, this allows you to modify the labels of the variable groups
which will appear in the facet strips, e.g. facet_var = grouping_variable,
facet_var_labs = c("group_1_new_label" = "group_1_old_label",
"group_2_new_label" = "group_2_old_label"). See
\code{\link[forcats]{fct_recode}} for details.}

\item{facet_var_strip_position}{If a variable has been assigned for facetting
using facet_var, this allows you to modify the position of the facet strip
labels. Sensible options include "top" (the default) or "bottom".}

\item{facet_var_text_bold}{If a variable has been assigned for facetting
using facet_var, this allows you to use boldface (TRUE/default or FALSE)
for the facet strip label text.}

\item{legend_position}{This allows you to modify the legend position.
Options include "right" (the default), "left", "top", & "bottom".}

\item{omit_legend}{Set to TRUE if you want to remove/omit the legends.}

\item{interactive}{Determines whether a static ggplot object or an interactive html
plotly object is returned. See \code{\link[plotly]{ggplotly}} for details.}

\item{aesthetic_options}{If set to TRUE, opens a web browser to the tidyverse
online aesthetic options vignette.}
}
\value{
A ggplot object or plotly object depending on whether static or
interactive output was requested.
}
\description{
Easily generate a density plot of a numeric variable using
ggplot2 with a simplified customization interface for common modifications
with static (ggplot) and interactive (plotly) output options. The static
output is useful for producing static reports (e.g. for manuscripts) and is
readily customized further using ggplot2 syntax. The interactive output is
helpful for exploring the data and producing dynamic html reports.
}
\examples{
data(mtcars) #load the mtcars data

library(magrittr)

mtcars \%>\% plot_density(x = mpg)

\dontrun{
mtcars \%>\%
  plot_density(x = mpg, transform_x = T, x_transformation = "log2")

mtcars \%>\%
  plot_density(x = mpg, transform_x = T, x_transformation = "log10") #default transformation

mtcars \%>\%
  plot_density(x = mpg, transform_x = T, x_transformation = "sqrt") #default transformation

mtcars \%>\% plot_density(x = mpg, fill_var = cyl, fill_var_labs = c("four" = "4"))

mtcars \%>\% plot_density(x = mpg, fill_var = cyl, fill_var_title = "# cyl",
                        interactive = T)

mtcars \%>\% plot_density(x = mpg,
                        fill_var = am,
                        fill_var_order = c("1", "0"),
                        fill_var_labs = c("manual" = "0",
                                          "automatic" = "1"),
                        fill_var_values = c("blue4", "red4"),
                        fill_var_title = "transmission")

mtcars \%>\% plot_density(x = mpg,
                        colour_var = cyl,
                        colour_var_order = c("6", "8", "4"),
                        colour_var_labs = c("six" = "6", "eight" = "8"),
                        colour_var_values = c("blue3", "red3", "green3"),
                        colour_var_title = "# cylinders")

#interactive version

mtcars \%>\% describe(mpg, fill_var = cyl, interactive = T)
}


}
\references{
Wickham, H. (2016). ggplot2: elegant graphics for data analysis. New York, N.Y.: Springer-Verlag.
}
\seealso{
\code{\link[ggplot2]{geom_density}}, \code{\link[plotly]{ggplotly}}
}
\author{
Craig P. Hutton, \email{Craig.Hutton@gov.bc.ca}
}
